Move all code from Data Layer to the Data Module
