+++ app/src/main/java/com/kennedydias/libermovies/LiberMoviesApplication.kt
@@ -1,8 +1,8 @@
 package com.kennedydias.libermovies
 
 import android.app.Application
-import com.kennedydias.libermovies.di.CommonModules
-import com.kennedydias.libermovies.di.RepositoryModules
+import com.kennedydias.data.di.CommonModules
+import com.kennedydias.data.di.RepositoryModules
 import com.kennedydias.libermovies.di.UseCases
 import com.kennedydias.libermovies.di.ViewModelModules
 import org.koin.core.context.startKoin
