+++ app/src/main/java/com/kennedydias/libermovies/ui/base/OnBackPressed.kt
@@ -1,4 +1,4 @@
-package com.kennedydias.libermovies.base
+package com.kennedydias.libermovies.ui.base
 
 interface OnBackPressed {
     /**
